+++ node/src/test/kotlin/net/corda/node/services/MockServiceHubInternal.kt
@@ -12,7 +12,6 @@ import net.corda.node.serialization.NodeClock
 import net.corda.node.services.api.*
 import net.corda.node.services.messaging.MessagingService
 import net.corda.node.services.schema.NodeSchemaService
-import net.corda.node.services.statemachine.FlowLogicRefFactoryImpl
 import net.corda.node.services.statemachine.FlowStateMachineImpl
 import net.corda.node.services.statemachine.StateMachineManager
 import net.corda.node.services.transactions.InMemoryTransactionVerifierService
@@ -30,7 +29,6 @@ open class MockServiceHubInternal(
         val mapCache: NetworkMapCacheInternal? = MockNetworkMapCache(),
         val scheduler: SchedulerService? = null,
         val overrideClock: Clock? = NodeClock(),
-        val flowFactory: FlowLogicRefFactoryInternal? = FlowLogicRefFactoryImpl(),
         val schemas: SchemaService? = NodeSchemaService(),
         val customTransactionVerifierService: TransactionVerifierService? = InMemoryTransactionVerifierService(2)
 ) : ServiceHubInternal() {
@@ -56,8 +54,8 @@ open class MockServiceHubInternal(
         get() = throw UnsupportedOperationException()
 
     override val monitoringService: MonitoringService = MonitoringService(MetricRegistry())
-    override val flowLogicRefFactory: FlowLogicRefFactoryInternal
-        get() = flowFactory ?: throw UnsupportedOperationException()
+    override val rpcFlows: List<Class<out FlowLogic<*>>>
+        get() = throw UnsupportedOperationException()
     override val schemaService: SchemaService
         get() = schemas ?: throw UnsupportedOperationException()
     override val auditService: AuditService = DummyAuditService()